% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{cbrt}
\alias{cbrt}
\title{Defines numeric types}
\usage{
cbrt(x)
}
\arguments{
\item{x}{numeric}
}
\value{
Integer with the maxnumber of threads
}
\description{
function to return x^1/3
}
\examples{
{
  cbrt(1:3)
}
}
