% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{dp_lineq_gausselim}
\alias{dp_lineq_gausselim}
\title{gaussian lineq}
\usage{
dp_lineq_gausselim(a, b)
}
\arguments{
\item{a}{matrix}

\item{b}{numeric}
}
\description{
gaussian lineq
}
\examples{
{
  ma <- array(1:6, c(2, 3))
  vb <- 1:5
  dp_lineq_gausselim(ma, vb)
}
}
