% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{dp_quadratic_pascal}
\alias{dp_quadratic_pascal}
\title{quadratic}
\usage{
dp_quadratic_pascal(bb, cc, a)
}
\arguments{
\item{bb}{numeric}

\item{cc}{numeric}

\item{a}{numericL}
}
\description{
quadratic
}
\examples{
{
  dp_quadratic_pascal(1:4, 4:1, 4:1)
}
}
