% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{dp_quadratic_pascal_reduced}
\alias{dp_quadratic_pascal_reduced}
\title{quadratic}
\usage{
dp_quadratic_pascal_reduced(bb, cc, a)

dp_quadratic_pascal_reduced(bb, cc, a)
}
\arguments{
\item{bb}{numeric}

\item{cc}{numeric}

\item{a}{numericL}
}
\description{
quadratic

quadratic
}
\examples{
{
  dp_quadratic_pascal_reduced(1:4, 4:1, 4:1)
}
{
  dp_quadratic_pascal_reduced(1:4, 4:1, 4:1)
}
}
