% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{dp_quadratic_reduced}
\alias{dp_quadratic_reduced}
\title{quadratic reduced}
\usage{
dp_quadratic_reduced(bb, cc)
}
\arguments{
\item{bb}{numeric}

\item{cc}{numeric}
}
\description{
quadratic reduced
}
\examples{
{
  dp_quadratic_reduced(1:4, 4:1)
}
}
