% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{get_threads}
\alias{get_threads}
\title{Check the number of threads}
\usage{
get_threads()
}
\value{
Integer with the maxnumber of threads
}
\description{
\code{get_threads} check the number of threads in this machine
}
\examples{
{
  get_threads()
}
}
