% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{get_types}
\alias{get_types}
\title{Defines numeric types}
\usage{
get_types()
}
\value{
Integer with the maxnumber of threads
}
\description{
\code{get_threads} check the number of threads in this machine
}
\examples{
{
get_threads()
}
}
