% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{quadratic}
\alias{quadratic}
\title{quadratic}
\usage{
quadratic(bb, cc, a)
}
\arguments{
\item{bb}{numeric}

\item{cc}{numeric}

\item{a}{numericL}
}
\description{
quadratic
}
\examples{
{
quadratic(1:4,4:1,4:1)
}
}
