% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{quadratic_reduced}
\alias{quadratic_reduced}
\title{quadratic reduced}
\usage{
quadratic_reduced(bb, cc)
}
\arguments{
\item{bb}{numeric}

\item{cc}{numeric}
}
\description{
quadratic reduced
}
\examples{
{
quadratic_reduced(1:4,4:1)
}
}
